Add bid and budget snapshots
Once bid and budget was collected from Ads API we can re-use it to
define snapshots - tables that contain bid and budget values for a
particular date. Based on those we can easily identify the values of
budget on a given day (which is quite complex to do with Ads API).

Change-Id: Ic2407d36d6ff449d454b50f2e313435bcd41616a